Add support for arbitrary clock multipliers
Allows users to set an arbitrary floating point value to multiply CPU clock speed by. Drop box provides access to all the previous presets.

+++ libpcsxcore/psxcommon.h
@@ -166,7 +166,7 @@ typedef struct {
 	u8 HackFix;
 	u8 MemHack;
 	boolean OverClock;	// enable overclocking
-	u8 PsxClock;
+	float PsxClock;
 	// PGXP variables
 	boolean PGXP_GTE;
 	boolean PGXP_Cache;
